requirements.txt
# Core dependencies
mcp
atlassian-python-api
beautifulsoup4

# Testing dependencies
pytest
pytest-asyncio

# Optional: for better HTML parsing
lxml
